requirements.txt
asgiref==3.8.1 
Django==4.2.13 
djangorestframework==3.15.1 
psycopg2-binary
python-dateutil==2.9.0.post0 
six==1.16.0 
sqlparse==0.5.0 
typing_extensions==4.11.0 
tzdata==2024.1
django-colorfield==0.11.0
gurobipy==11.0.0